# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: URL Shortener API
  version: 1.0.0
  description: API runtime contract for owned link management, redirects, click totals, system information, and health.
servers:
  - url: http://localhost:8080
    description: Local Docker Compose
...
paths:
...
  /api/v1/links/{code}/analytics:
    get:
      tags: [Analytics]
      summary: Get the total persisted clicks for a short link
      description: Requires the capability token returned once when the link is created.
      operationId: getLinkAnalytics
      security:
        - OwnerToken: []
      parameters:
        - $ref: '#/components/parameters/Code'
      responses:
        '200':
          description: Click total
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          $ref: '#/components/responses/Gone'
...
components:
  securitySchemes:
    OwnerToken:
      type: apiKey
      in: header
      name: X-Link-Owner-Token
  parameters:
    Code:
      name: code
      in: path
      required: true
      schema:
        type: string
        pattern: '^[A-Za-z0-9_-]{4,32}$'
  responses:
...
    Forbidden:
      description: Missing or invalid owner token
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    NotFound:
      description: Short link not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Gone:
      description: Short link expired or disabled
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    AnalyticsResponse:
      type: object
      required: [code, totalClicks]
      properties:
        code:
          type: string
        totalClicks:
          type: integer
          format: int64
          minimum: 0
...
    Problem:
      type: object
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri
        errors:
          type: array
          items:
            type: string
      additionalProperties: true